requirements.txt
pillow
paho-mqtt
websocket-client
numpy
rel